backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
motor>=3.4.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
pywebpush>=2.0.0
pydantic>=2.0.0
